fix: use 32bit integer for pn/dn

+++ source/engine/user-engine/typedefs.hpp
@@ -21,8 +21,8 @@ inline constexpr Key kNullKey = Key{0};
 template <bool kOrNode>
 struct NodeTag {};
 
-/// 証明数／反証数を格納する型。将来、(1,1) 以外の初期値を使うことを考慮して 64 bit 分確保する。
-using PnDn = std::uint64_t;
+/// 証明数／反証数を格納する型
+using PnDn = std::uint32_t;
 /// pn/dn の最大値。オーバーフローを避けるために、max() より少し小さな値を設定する。
 inline constexpr PnDn kInfinitePnDn = std::numeric_limits<PnDn>::max() / 2;
